test: try splitting out tests
