+++ .github/workflows/ci.yml
@@ -37,6 +37,9 @@ jobs:
       - name: Run tests
         run: npm test
 
+      - name: Run tests (single-threaded)
+        run: npm test:single
+
   lint:
     runs-on: ubuntu-latest
