+++ __tests__/cmds/openapi/convert.test.ts
@@ -1,7 +1,7 @@
 import fs from 'fs';
 
 import prompts from 'prompts';
-import { describe, afterEach, it, expect, vi } from 'vitest';
+import { describe, it, expect, vi } from 'vitest';
 
 import OpenAPIConvertCommand from '../../../src/cmds/openapi/convert';
 
@@ -9,15 +9,7 @@ const convert = new OpenAPIConvertCommand();
 
 const successfulConversion = () => 'Your converted API definition has been saved to output.json!';
 
-const testWorkingDir = process.cwd();
-
 describe('rdme openapi:convert', () => {
-  afterEach(() => {
-    process.chdir(testWorkingDir);
-
-    vi.clearAllMocks();
-  });
-
   describe('converting', () => {
     it.each([
       ['Swagger 2.0', 'json', '2.0'],
@@ -40,27 +32,6 @@ describe('rdme openapi:convert', () => {
 
       expect(fs.writeFileSync).toHaveBeenCalledWith('output.json', expect.any(String));
       expect(reducedSpec.tags).toHaveLength(1);
-      expect(Object.keys(reducedSpec.paths)).toStrictEqual(['/pet/{petId}']);
-      expect(Object.keys(reducedSpec.paths['/pet/{petId}'])).toStrictEqual(['get', 'post', 'delete']);
-    });
-
-    it.skip('should convert with no prompts via opts', async () => {
-      const spec = 'petstore-simple.json';
-
-      let reducedSpec;
-      fs.writeFileSync = vi.fn((fileName, data) => {
-        reducedSpec = JSON.parse(data as string);
-      });
-
-      await expect(
-        convert.run({
-          spec,
-          workingDirectory: require.resolve(`@readme/oas-examples/2.0/json/${spec}`).replace(spec, ''),
-          out: 'output.json',
-        }),
-      ).resolves.toBe(successfulConversion());
-
-      expect(fs.writeFileSync).toHaveBeenCalledWith('output.json', expect.any(String));
       expect(Object.keys(reducedSpec.paths)).toStrictEqual(['/pet/{petId}']);
       expect(Object.keys(reducedSpec.paths['/pet/{petId}'])).toStrictEqual(['get', 'post', 'delete']);
     });
