+++ __tests__/cmds/openapi/index.test.ts
@@ -45,8 +45,6 @@ const successfulUpdate = (specPath, specType = 'OpenAPI') =>
     ...successfulMessageBase(specPath, specType),
   ].join('\n');
 
-const testWorkingDir = process.cwd();
-
 const getCommandOutput = () => {
   return [consoleWarnSpy.mock.calls.join('\n\n'), consoleInfoSpy.mock.calls.join('\n\n')].filter(Boolean).join('\n\n');
 };
@@ -68,8 +66,6 @@ describe('rdme openapi', () => {
     consoleWarnSpy.mockRestore();
 
     nock.cleanAll();
-
-    process.chdir(testWorkingDir);
   });
 
   describe('upload', () => {
@@ -120,44 +116,6 @@ describe('rdme openapi', () => {
       expect(console.info).toHaveBeenCalledTimes(0);
 
       postMock.done();
-      return mock.done();
-    });
-
-    it.skip('should discover and upload an API definition if none is provided', async () => {
-      const registryUUID = getRandomRegistryId();
-
-      const mock = getAPIMock()
-        .get(`/api/v1/version/${version}`)
-        .basicAuth({ user: key })
-        .reply(200, { version })
-        .post('/api/v1/api-registry', body => body.match('form-data; name="spec"'))
-        .reply(201, { registryUUID, spec: { openapi: '3.0.0' } });
-
-      const mockWithHeader = getAPIMockWithVersionHeader(version)
-        .get('/api/v1/api-specification')
-        .basicAuth({ user: key })
-        .reply(200, [])
-        .post('/api/v1/api-specification', { registryUUID })
-        .delayConnection(1000)
-        .basicAuth({ user: key })
-        .reply(201, { _id: 1 }, { location: exampleRefLocation });
-
-      const spec = 'petstore.json';
-
-      await expect(
-        openapi.run({
-          key,
-          version,
-          workingDirectory: './__tests__/__fixtures__/relative-ref-oas',
-        }),
-      ).resolves.toBe(successfulUpload(spec));
-
-      expect(console.info).toHaveBeenCalledTimes(1);
-
-      const output = getCommandOutput();
-      expect(output).toBe(chalk.yellow(`ℹ️  We found ${spec} and are attempting to upload it.`));
-
-      mockWithHeader.done();
       return mock.done();
     });
 
@@ -372,83 +330,6 @@ describe('rdme openapi', () => {
       `);
 
       postMock.done();
-      return mock.done();
-    });
-
-    it.skip('should use specified working directory and upload the expected content', async () => {
-      let requestBody;
-      const registryUUID = getRandomRegistryId();
-      const mock = getAPIMock()
-        .get(`/api/v1/version/${version}`)
-        .basicAuth({ user: key })
-        .reply(200, { version: '1.0.0' })
-        .post('/api/v1/api-registry', body => {
-          requestBody = body.substring(body.indexOf('{'), body.lastIndexOf('}') + 1);
-          requestBody = JSON.parse(requestBody);
-
-          return body.match('form-data; name="spec"');
-        })
-        .reply(201, { registryUUID, spec: { openapi: '3.0.0' } });
-
-      const mockWithHeader = getAPIMockWithVersionHeader(version)
-        .get('/api/v1/api-specification')
-        .basicAuth({ user: key })
-        .reply(200, [])
-        .post('/api/v1/api-specification', { registryUUID })
-        .basicAuth({ user: key })
-        .reply(201, { _id: 1 }, { location: exampleRefLocation });
-
-      const spec = 'petstore.json';
-
-      await expect(
-        openapi.run({
-          spec,
-          key,
-          version,
-          workingDirectory: './__tests__/__fixtures__/relative-ref-oas',
-        }),
-      ).resolves.toBe(successfulUpload(spec));
-
-      expect(console.info).toHaveBeenCalledTimes(0);
-
-      expect(requestBody).toMatchSnapshot();
-
-      mockWithHeader.done();
-      return mock.done();
-    });
-
-    it.skip('should return spec create info for dry run', async () => {
-      const registryUUID = getRandomRegistryId();
-
-      const mock = getAPIMock()
-        .get(`/api/v1/version/${version}`)
-        .basicAuth({ user: key })
-        .reply(200, { version })
-        .post('/api/v1/api-registry', body => body.match('form-data; name="spec"'))
-        .reply(201, { registryUUID, spec: { openapi: '3.0.0' } });
-
-      const mockWithHeader = getAPIMockWithVersionHeader(version)
-        .get('/api/v1/api-specification')
-        .basicAuth({ user: key })
-        .reply(200, []);
-
-      await expect(
-        openapi.run({
-          key,
-          version,
-          dryRun: true,
-          workingDirectory: './__tests__/__fixtures__/relative-ref-oas',
-        }),
-      ).resolves.toMatch(
-        '🎭 dry run! The API Definition located at petstore.json will be created for this project version: 1.0.0',
-      );
-
-      const output = getCommandOutput();
-      expect(output).toMatch(
-        chalk.yellow('🎭 dry run option detected! No API definitions will be created or updated in ReadMe.'),
-      );
-
-      mockWithHeader.done();
       return mock.done();
     });
   });
@@ -900,14 +781,6 @@ describe('rdme openapi', () => {
       ).rejects.toStrictEqual(new APIError(errorObject));
 
       return mock.done();
-    });
-
-    it.skip('should error if no file was provided or able to be discovered', () => {
-      return expect(openapi.run({ key, version, workingDirectory: 'config' })).rejects.toStrictEqual(
-        new Error(
-          "We couldn't find an OpenAPI or Swagger definition.\n\nPlease specify the path to your definition with `rdme openapi ./path/to/api/definition`.",
-        ),
-      );
     });
 
     it('should throw an error if an invalid OpenAPI 3.0 definition is supplied', () => {
@@ -1380,46 +1253,6 @@ describe('rdme openapi', () => {
       return mock.done();
     });
 
-    it.skip('should create GHA workflow (including workingDirectory)', async () => {
-      expect.assertions(4);
-      const yamlFileName = 'openapi-file-workingdirectory';
-      prompts.inject([true, 'openapi-branch-workingdirectory', yamlFileName]);
-      const registryUUID = getRandomRegistryId();
-
-      const mock = getAPIMock()
-        .get(`/api/v1/version/${version}`)
-        .basicAuth({ user: key })
-        .reply(200, { version: '1.0.0' })
-        .post('/api/v1/api-registry', body => body.match('form-data; name="spec"'))
-        .reply(201, { registryUUID, spec: { openapi: '3.0.0' } });
-
-      const mockWithHeader = getAPIMockWithVersionHeader(version)
-        .get('/api/v1/api-specification')
-        .basicAuth({ user: key })
-        .reply(200, [])
-        .post('/api/v1/api-specification', { registryUUID })
-        .basicAuth({ user: key })
-        .reply(201, { _id: 1 }, { location: exampleRefLocation });
-
-      const spec = 'petstore.json';
-
-      await expect(
-        openapi.run({
-          spec,
-          key,
-          version,
-          workingDirectory: './__tests__/__fixtures__/relative-ref-oas',
-        }),
-      ).resolves.toMatchSnapshot();
-
-      expect(yamlOutput).toMatchSnapshot();
-      expect(fs.writeFileSync).toHaveBeenCalledTimes(2);
-      expect(fs.writeFileSync).toHaveBeenNthCalledWith(2, `.github/workflows/${yamlFileName}.yml`, expect.any(String));
-
-      mockWithHeader.done();
-      return mock.done();
-    });
-
     it('should reject if user says no to creating GHA workflow', async () => {
       prompts.inject(['create', false]);
       const registryUUID = getRandomRegistryId();
@@ -1546,48 +1379,6 @@ describe('rdme openapi', () => {
       putMock.done();
       exampleMock.done();
       return mock.done();
-    });
-
-    it.skip('should contain request header with correct URL with working directory', async () => {
-      const registryUUID = getRandomRegistryId();
-      const mock = getAPIMock()
-        .get(`/api/v1/version/${version}`)
-        .basicAuth({ user: key })
-        .reply(200, { version: '1.0.0' })
-        .post('/api/v1/api-registry')
-        .reply(201, { registryUUID, spec: { openapi: '3.0.0' } });
-
-      const getMock = getAPIMockWithVersionHeader(version)
-        .get('/api/v1/api-specification')
-        .basicAuth({ user: key })
-        .reply(200, []);
-
-      const postMock = getAPIMock({
-        'x-rdme-ci': 'GitHub Actions (test)',
-        'x-readme-source': 'cli-gh',
-        'x-readme-source-url':
-          'https://github.com/octocat/Hello-World/blob/ffac537e6cbbf934b08745a378932722df287a53/__tests__/__fixtures__/relative-ref-oas/petstore.json',
-        'x-readme-version': version,
-      })
-        .post('/api/v1/api-specification', { registryUUID })
-        .basicAuth({ user: key })
-        .reply(201, { _id: 1 }, { location: exampleRefLocation });
-
-      const spec = 'petstore.json';
-
-      await expect(
-        openapi.run({
-          spec,
-          key,
-          version,
-          workingDirectory: './__tests__/__fixtures__/relative-ref-oas',
-        }),
-      ).resolves.toBe(successfulUpload(spec));
-
-      getMock.done();
-      postMock.done();
-      mock.done();
-      return after();
     });
   });
 });
